requirements.txt
streamlit
requests
google-generativeai
chromadb
python-dotenv
pyngrok
